unsigned long -->> int

+++ lwext4/ext4_extent.h
@@ -232,10 +232,10 @@ static inline void ext4_idx_store_pblock(struct ext4_extent_idx *ix,
 
 int ext4_ext_get_blocks(struct ext4_inode_ref *inode_ref,
 			ext4_fsblk_t iblock,
-			unsigned long max_blocks,
+			int max_blocks,
 			ext4_fsblk_t *result,
 			int create,
-			unsigned long *blocks_count);
+			int *blocks_count);
 
 int ext4_ext_tree_init(struct ext4_inode_ref *inode_ref);
